backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyMuPDF>=1.23.0
openai>=1.0.0
python-dotenv>=1.0.0
unstructured[pdf]>=0.10.0
